Implemented interval option, better defaults and a bug fix.
Interval option built with a recursive setTimeout (not setInterval) that tracks it's invoked number.

Patch function now determines it's own ID if it isn't provided, meaning that the opts-must-be-object requirement isn't needed. In other words, all options are optional (wow).

Bug fix: minor thing where I was using gun before it was defined. This was due to minor tweaking just before a commit and not having unit tests to shout at me :(