Add async done events, multi-test handlers, test run/end methods, client managers, tests failing.
Each test now tracks each peer that runs it, and can dig into the server clients to see if they're still connected. As clients report that they're done, it trickles into the test and it'll match it against the number of registered runners to see if it's all finished.

The stack now has a primitive method of pushing out the next test when it's ready, so you can write multiple tests and they'll run one after the other.

Tests now have a run and end method that allow you to easily tell the test when is appropriate to begin and when it should officially end and send control to the next test.

The tests are now notified when clients join and finish.

Sadly, the tests are still failing :(