Filter clients on the server, save test results, isolate test controls.
Platform details are now sent to the server. If there are no stage listeners, it automatically sends the test to all the connected clients without being picky. This will allow extensions to filter out clients that don't match the criteria.

The results of every test are saved into a list of responses, each scoped to their test.

Instead of the stack deciding which clients get the test, that's been delegated to the test instances.