Merge gather with stage event, fix nomenclature, remove reporters
The gather method is now reactive with the "stage" event, so it won't execute and prematurely begin the test. The "needs" method now uses gather internally and decides when to `run`.

The "peer-done" event didn't make much sense for a centralized configuration, so it's been changed to "client-done".

Since the reporters have been moved out into the panic-cli repo, they've been removed from panic-server.