Remove notes, serialize functions properly, add data method.
Notes don't belong in core. Not sure why I included them :(

Functions are now toString'd when toJSON is called, allowing them to be
sent over the wire.

The new data method for clients is meant to set environment variables.
Send in the name of the scope and it will be merged in. This API may
change later to allow brevity.