Add scope controls, loosely type check platform queries.
All tests passing.

Added a new keyword (cross-platform) to the package.json file.

Optionally, scope variables can be sent to the clients in the `.run`
method.

Platform queries are now loosely type checked. This allows you to match
a number against a string (or vice versa), such as a version or
architecture.